dtype

+++ pandas/tests/test_base.py
@@ -1236,7 +1236,7 @@ def test_values_consistent(array, expected_type, dtype):
      np.array(['2017-01-01T05:00:00'], dtype='M8[ns]')),
     (pd.TimedeltaIndex([10**10]), np.array([10**10], dtype='m8[ns]')),
     (pd.PeriodIndex(['2017', '2018'], freq='D'),
-     np.array([17167, 17532])),
+     np.array([17167, 17532], dtype=np.int64)),
 ])
 def test_ndarray_values(array, expected):
     l_values = pd.Series(array)._ndarray_values
